pyproject.toml
[tool.poetry]
name = "yetien-asr-model-server-grpc"
version = "0.1.0"
description = ""
authors = ["Jianxuan Li"]
readme = "README.md"
packages = [{include = "server"}]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.56.0"
grpcio-tools = "^1.56.0"
redis = "^4.6.0"
boto3 = "^1.28.3"
requests = "^2.31.0"
coverage = "^7.2.7"
protobuf = "^4.23.4"


[tool.poetry.group.dev.dependencies]
pycodestyle = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
